+++ django_rest_passwordreset/signals.py
@@ -12,11 +12,11 @@ Signal arguments: instance, reset_password_token
 reset_password_token_created = Signal()
 
 """
-Signal arguments: user
+Signal arguments: user, reset_password_token
 """
 pre_password_reset = Signal()
 
 """
-Signal arguments: user
+Signal arguments: user, reset_password_token
 """
 post_password_reset = Signal()