+++ django_rest_passwordreset/views.py
@@ -82,7 +82,11 @@ class ResetPasswordConfirm(GenericAPIView):
 
         # change users password (if we got to this code it means that the user is_active)
         if reset_password_token.user.eligible_for_reset():
-            pre_password_reset.send(sender=self.__class__, user=reset_password_token.user)
+            pre_password_reset.send(
+                sender=self.__class__,
+                user=reset_password_token.user,
+                reset_password_token=reset_password_token,
+            )
             try:
                 # validate the password against existing validators
                 validate_password(
@@ -98,7 +102,11 @@ class ResetPasswordConfirm(GenericAPIView):
 
             reset_password_token.user.set_password(password)
             reset_password_token.user.save()
-            post_password_reset.send(sender=self.__class__, user=reset_password_token.user)
+            post_password_reset.send(
+                sender=self.__class__,
+                user=reset_password_token.user,
+                reset_password_token=reset_password_token,
+            )
 
         # Delete all password reset tokens for this user
         ResetPasswordToken.objects.filter(user=reset_password_token.user).delete()